Make setting.html and move theme change to setting.
Make setting.html to manage functions like theme change and others.
