+++ Codify/css/codify/style_black.css
@@ -82,6 +82,10 @@ select#lang {
     margin-left: -0.8vw;
 }
 
+#selected-theme {
+    color: #0f0f0f;
+}
+
 .onoffswitch {
     position: relative; width: 47px; top: 2px;
     -webkit-user-select:none; -moz-user-select:none; -ms-user-select: none;
