+++ Codify/html/result.html
@@ -47,8 +47,7 @@
             <li class="active"><a id="a-editor" href="#">Editor</a></li>
             <li><a id="a-selector" href="#">Selector</a></li>
             <li><a id="a-logger" href="#">Logger</a></li>
-            <li><a id="a-change-theme">Change Theme</a></li>
-            <li><a href="#">Setting</a></li>
+            <li><a id="a-setting">Setting</a></li>
           </ul>
         </div>
       </div>
