+++ Codify/html/setting.html
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+<html lang="ko">
+
+<head>
+    <meta charset="utf-8" />
+    <!-- bootstrap -->
+    <link rel="stylesheet" type="text/css" href="/css/bootstrap.min.css">
+    <link rel="stylesheet" type="text/css" href="/css/bootstrap-theme.min.css">
+    <script src="/js/jquery-3.4.1.min.js"></script>
+    <script src="/js/bootstrap.min.js"></script>
+    <!-- font awesome -->
+    <link rel="stylesheet" type="text/css" href="/css/all.min.css">
+    <!-- custom -->
+    <link rel="stylesheet" type="text/css" href="/css/codify/all.css">
+    <link rel="stylesheet" id="theme" type="text/css" href="/css/codify/style_black.css">
+</head>
+
+<body>
+<span class="titlebar">
+    Codify ver.1.0
+  </span>
+
+<nav class="navbar navbar-inverse">
+    <div class="container-fluid">
+        <div class="navbar-header">
+            <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#myNavbar">
+                <span class="icon-bar"></span>
+                <span class="icon-bar"></span>
+                <span class="icon-bar"></span>
+            </button>
+            <span id="navbar-onoff" class="navbar-brand">
+          <div class="onoffswitch">
+            <input type="checkbox" name="onoffswitch" class="onoffswitch-checkbox" id="myonoffswitch" checked>
+            <label class="onoffswitch-label" for="myonoffswitch">
+              <span class="onoffswitch-inner"></span>
+              <span class="onoffswitch-switch"></span>
+            </label>
+          </div>
+        </span>
+        </div>
+        <div class="collapse navbar-collapse" id="myNavbar">
+            <ul class="nav navbar-nav">
+                <li><a id="a-editor" href="#">Editor</a></li>
+                <li><a id="a-selector" href="#">Selector</a></li>
+                <li><a id="a-logger" href="#">Logger</a></li>
+                <li class="active"><a id="a-setting">Setting</a></li>
+            </ul>
+        </div>
+    </div>
+</nav>
+
+<div id="setting-section">
+    <label class="section-label">SETTING</label>
+    <label>Editor Font Size: </label>
+    <table class="table table-dark table-hover">
+        <thead>
+        <tr>
+            <th style="width: 40px;">name</th>
+            <th style="width: 120px;">value</th>
+        </tr>
+        </thead>
+        <tbody>
+        <tr>
+            <th style="width: 40px;">Change Theme</th>
+            <th style="width: 120px;">
+                <!-- Could add new theme to here -->
+                <select id="selected-theme" class="change-theme" size="1">
+                    <option id="" selected>Choose Theme Color</option>
+                    <option id="black">Black</option>
+                    <option id="white">White</option>
+                </select>
+            </th>
+        </tr>
+        </tbody>
+    </table>
+</div>
+
+<script type="module" src="/js/codify/html_all.js"></script>
+<script type="module" src="/js/codify/settings.js"></script>
+<script type="module" src="/js/codify/nav.js"></script>
+</body>
+
+</html>
